Pull plugins into their own package, plugin-utils
- Will be used by embed-grid and embed-chart as well
