+++ packages/code-studio/package.json
@@ -29,6 +29,7 @@
     "@deephaven/jsapi-types": "file:../jsapi-types",
     "@deephaven/jsapi-utils": "file:../jsapi-utils",
     "@deephaven/log": "file:../log",
+    "@deephaven/plugin-utils": "file:../plugin-utils",
     "@deephaven/pouch-storage": "file:../pouch-storage",
     "@deephaven/react-hooks": "file:../react-hooks",
     "@deephaven/redux": "file:../redux",
@@ -36,8 +37,6 @@
     "@deephaven/utils": "file:../utils",
     "@fortawesome/fontawesome-svg-core": "^6.2.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@paciolan/remote-component": "2.7.2",
-    "@paciolan/remote-module-loader": "^3.0.2",
     "classnames": "^2.3.1",
     "event-target-shim": "^6.0.2",
     "fira": "mozilla/fira#4.202",
