+++ packages/code-studio/tsconfig.json
@@ -58,6 +58,9 @@
       "path": "../log"
     },
     {
+      "path": "../plugin-utils"
+    },
+    {
       "path": "../pouch-storage"
     },
     {
