+++ packages/plugin-utils/README.md
@@ -1,0 +1,11 @@
+# @deephaven/plugin-utils
+
+A library with some utility functions used by Deephaven to load and manage plugins.
+
+For more details on using Deephaven JS Plugins, see the [deephaven-js-plugins repository](https://github.com/deephaven/deephaven-js-plugins).
+
+## Install
+
+```bash
+npm install --save @deephaven/plugin-utils
+```
