+++ packages/plugin-utils/src/index.ts
@@ -1,2 +1,2 @@
-export { default as PluginUtils } from './PluginUtils';
+export * from './PluginUtils';
 export { default as RemoteComponent } from './RemoteComponent';
