+++ packages/plugin-utils/src/loadRemoteModule.ts
@@ -1,7 +1,7 @@
 import createLoadRemoteModule, {
   createRequires,
 } from '@paciolan/remote-module-loader';
-import { resolve } from '../remote-component.config';
+import { resolve } from './remote-component.config';
 
 const requires = createRequires(resolve);
