+++ packages/plugin-utils/src/remote-component.config.js
